added confirmation for clearing the entire error cache

+++ src/Directives/settingsModal.js
@@ -85,6 +85,7 @@ export default function(ngapp, fileHelpers) {
         };
 
         $scope.clearErrorCache = function() {
+            if (!confirm('Clear the entire error cache?')) return;
             $scope.errorCache.forEach(function(cache) {
                 $scope.deleteCache(cache);
             });
